refactor(moq-video): move the rate control policy out of moq-net
The policy is an encoder concern wearing a u64 as a disguise: its floor is
justified by the picture not being worth sending, its headroom by leaving room
for audio, its ramp by a contribution encoder on a cellular uplink. None of
that is generic, and moq-net is the media-agnostic layer.

Nothing in moq-net used it either, so it was permanent public surface on a
published crate for a single external caller, and it collided by name with
ietf::Control. It lives in moq-video::encode::rate now, whose pub API is free
to churn. If the audio budget ever needs it, hang is the shared home to
promote it to, with two real consumers to design against.

bandwidth::Producer/Consumer stay in moq-net: the estimate is transport, only
what a sender does with it is policy.
